refactor: fold ExnMessageTypes in ExnMessages
