+++ src/main/java/org/cardanofoundation/signify/app/Exchanging.java
@@ -177,16 +177,16 @@ public class Exchanging {
          * found or the route is unknown. Use the route-specific getters when the
          * expected type is already known.
          */
-        public Optional<ExnMessageTypes.TypedExchange> getTyped(String said) throws Exception {
-            return get(said).flatMap(ExnMessageTypes::asTyped);
+        public Optional<ExnMessages.TypedExchange> getTyped(String said) throws Exception {
+            return get(said).flatMap(ExnMessages::asTyped);
         }
 
         /**
          * Fetches an exchange message as the given typed form; empty when not found
          * or when its route does not produce that type.
          */
-        public <T extends ExnMessageTypes.TypedExchange> Optional<T> get(String said, Class<T> type) throws Exception {
-            return get(said).flatMap(msg -> ExnMessageTypes.as(msg, type));
+        public <T extends ExnMessages.TypedExchange> Optional<T> get(String said, Class<T> type) throws Exception {
+            return get(said).flatMap(msg -> ExnMessages.as(msg, type));
         }
 
     }
