+++ src/test/java/org/cardanofoundation/signify/e2e/utils/MultisigUtils.java
@@ -5,10 +5,10 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import org.cardanofoundation.signify.app.Exchanging;
-import org.cardanofoundation.signify.app.ExnMessageTypes.MultisigIcpExchange;
-import org.cardanofoundation.signify.app.ExnMessageTypes.MultisigIxnExchange;
+import org.cardanofoundation.signify.app.ExnMessages.MultisigIcpExchange;
+import org.cardanofoundation.signify.app.ExnMessages.MultisigIxnExchange;
 import static org.cardanofoundation.signify.app.ExnMessages.*;
-import static org.cardanofoundation.signify.app.ExnMessageTypes.as;
+import static org.cardanofoundation.signify.app.ExnMessages.as;
 import org.cardanofoundation.signify.app.aiding.CreateIdentifierArgs;
 import org.cardanofoundation.signify.app.aiding.RotateIdentifierArgs;
 import org.cardanofoundation.signify.app.clienting.SignifyClient;
